+++ tools/simple-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @simple-data-open/simple-cli
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @simple-data-open/utils@0.0.13
+
 ## 0.0.12
 
 ### Patch Changes